Create EllipticCurve class
This class is still a work in progress, right now it is able to add two
points on the curve, but it is not addressing many of the edge cases
that it will need to address. For example if the points should add up
to the infinite point, we do not currenly handle it.

As a side note I don't normally like so much magic in my tests, but I
think this is a good time to DRY it up.